giving addition if product want to be updated but it doesn't exist on updateProduct class

+++ src/main/java/pws/a/learningRESTful/ProductServiceController.java
@@ -51,6 +51,8 @@ return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
  
  @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
  public  ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product){
+     if(!productRepo.containsKey(id))
+         return new ResponseEntity<>("Product is not found", HttpStatus.NOT_FOUND);
      productRepo.remove(id);
      product.setId(id);
      productRepo.put(id, product);
